CONN-10476 Extract common part of SSv1 and SSv2
